Separate make commands that require sudo Reworked privileged cert test so it runs in Go by moving make logic to Go logic
