+++ pkg/storage/storage_fs_test.go
@@ -8,7 +8,6 @@ import (
 	"io/ioutil"
 	"math/big"
 	"os"
-	"os/exec"
 	"path"
 	"strings"
 	"testing"
@@ -598,79 +597,6 @@ func TestNegativeCases(t *testing.T) {
 
 		_, err = imgStore.PutBlobChunk("test", upload, 0, int64(l), buf)
 		So(err, ShouldNotBeNil)
-	})
-
-	Convey("Invalid dedupe scenarios", t, func() {
-		dir := t.TempDir()
-
-		log := log.Logger{Logger: zerolog.New(os.Stdout)}
-		metrics := monitoring.NewMetricsServer(false, log)
-		imgStore := storage.NewImageStore(dir, true, storage.DefaultGCDelay, true, true, log, metrics)
-
-		upload, err := imgStore.NewBlobUpload("dedupe1")
-		So(err, ShouldBeNil)
-		So(upload, ShouldNotBeEmpty)
-
-		content := []byte("test-data3")
-		buf := bytes.NewBuffer(content)
-		buflen := buf.Len()
-		digest := godigest.FromBytes(content)
-		blob, err := imgStore.PutBlobChunkStreamed("dedupe1", upload, buf)
-		So(err, ShouldBeNil)
-		So(blob, ShouldEqual, buflen)
-
-		blobDigest1 := strings.Split(digest.String(), ":")[1]
-		So(blobDigest1, ShouldNotBeEmpty)
-
-		err = imgStore.FinishBlobUpload("dedupe1", upload, buf, digest.String())
-		So(err, ShouldBeNil)
-		So(blob, ShouldEqual, buflen)
-
-		// Create a file at the same place where FinishBlobUpload will create
-		err = imgStore.InitRepo("dedupe2")
-		So(err, ShouldBeNil)
-
-		err = os.MkdirAll(path.Join(dir, "dedupe2", "blobs/sha256"), 0o755)
-		if err != nil {
-			panic(err)
-		}
-
-		err = ioutil.WriteFile(path.Join(dir, "dedupe2", "blobs/sha256", blobDigest1), content, 0o755) // nolint: gosec
-		if err != nil {
-			panic(err)
-		}
-
-		upload, err = imgStore.NewBlobUpload("dedupe2")
-		So(err, ShouldBeNil)
-		So(upload, ShouldNotBeEmpty)
-
-		content = []byte("test-data3")
-		buf = bytes.NewBuffer(content)
-		buflen = buf.Len()
-		digest = godigest.FromBytes(content)
-		blob, err = imgStore.PutBlobChunkStreamed("dedupe2", upload, buf)
-		So(err, ShouldBeNil)
-		So(blob, ShouldEqual, buflen)
-
-		cmd := exec.Command("sudo", "chattr", "+i", path.Join(dir, "dedupe2", "blobs/sha256", blobDigest1)) // nolint: gosec
-		_, err = cmd.Output()
-		if err != nil {
-			panic(err)
-		}
-
-		err = imgStore.FinishBlobUpload("dedupe2", upload, buf, digest.String())
-		So(err, ShouldNotBeNil)
-		So(blob, ShouldEqual, buflen)
-
-		cmd = exec.Command("sudo", "chattr", "-i", path.Join(dir, "dedupe2", "blobs/sha256", blobDigest1)) // nolint: gosec
-		_, err = cmd.Output()
-		if err != nil {
-			panic(err)
-		}
-
-		err = imgStore.FinishBlobUpload("dedupe2", upload, buf, digest.String())
-		So(err, ShouldBeNil)
-		So(blob, ShouldEqual, buflen)
 	})
 
 	Convey("DirExists call with a filename as argument", t, func(c C) {
